Extend default inputs
