+++ src/Table/Table.js
@@ -1,36 +1,9 @@
-import React, { useRef, useEffect, useState, useMemo, useCallback, createContext, useContext } from 'react';
+import React, { useMemo } from 'react';
 import classNames from 'classnames';
-import get from 'lodash/get';
-import setIn from 'lodash/fp/set';
+import Cells, { defaultEditComponents } from './Cells';
 import { useTable, useKeyboard, TableContext } from './';
 import { useScroller, ScrollerContainer, ScrollerCell } from '../Scroller';
 import GridResizer from '../grid/GridResizer';
-
-const StringEditComponent = ({ column, rowIndex, value = '', onChange, onEditingCellChange, ...props }) => {
-  const rootRef = useRef();
-
-  useEffect(() => {
-    rootRef.current.focus();
-  }, []);
-
-  const handleChange = useCallback(event => {
-    onChange(event.target.value);
-  }, [onChange]);
-
-  return (
-    <ScrollerCell
-        ref={rootRef}
-        Component="input"
-        column={column}
-        {...props}
-        value={value}
-        onChange={handleChange} />
-  );
-};
-
-const defaultEditComponents = {
-  string: StringEditComponent
-};
 
 const Heading = ({ index, onColumnsChange, column, defaultColumnWidth, style }) => {
   return (
@@ -69,215 +42,6 @@ const Header = React.memo(props => {
           style={style} />
     )
   })
-});
-
-const CellsRowContext = createContext();
-
-const CellsRow = ({ children }) => {
-  const [hoverRow, onHoverRowChange] = useState();
-  return (
-    <CellsRowContext.Provider value={{ hoverRow, onHoverRowChange }}>
-      {children}
-    </CellsRowContext.Provider>
-  )
-};
-
-const Cell = ({ value, column, ...props }) => {
-  return (
-    <ScrollerCell column={column} {...props}>
-      {column.format ? column.format(value) : value}
-    </ScrollerCell>
-  )
-};
-
-const EditCellWrapper = ({
-  Component,
-  value: valueProp,
-  onChange,
-  rowIndex,
-  column,
-  onEditingCellChange,
-  onSelectedCellsChange,
-  totalRows,
-  totalColumns,
-  ...props
-}) => {
-  const [value, setValue] = useState(valueProp);
-
-  const submitChange = useCallback(() => {
-    onChange(sourceValue => {
-      const nextValue = setIn(`[${rowIndex}].${column.valuePath}`, value, sourceValue);
-      return nextValue;
-    });
-  }, [column.valuePath, rowIndex, value, onChange]);
-
-  const onBlur = useCallback(() => {
-    submitChange();
-    onEditingCellChange(undefined);
-  }, [onEditingCellChange, submitChange]);
-
-  const onKeyDown = useCallback(event => {
-    switch(event.key) {
-      case 'Escape':
-        onEditingCellChange(undefined);
-        break;
-      case 'Enter':
-        submitChange();
-        onEditingCellChange(undefined);
-        break;
-      case 'Tab':
-        submitChange();
-        onEditingCellChange(editingCell => {
-          let nextEditingCell;
-          if (editingCell.row === totalRows - 1 && editingCell.column === totalColumns - 1) {
-            // Adding new row
-            onChange(value => [...value, {}]);
-            nextEditingCell = { row: totalRows, column: 0 };
-          } else if (editingCell.column === totalColumns - 1) {
-            nextEditingCell = { row: editingCell.row + 1, column: 0 };
-          } else {
-            nextEditingCell = { row: editingCell.row, column: editingCell.column + 1 };
-          }
-          onSelectedCellsChange([nextEditingCell]);
-          return nextEditingCell;
-        });
-        break;
-      default:
-    }
-  }, [onChange, submitChange, onEditingCellChange, onSelectedCellsChange, totalRows, totalColumns]);
-
-  return (
-    <Component
-        rowIndex={rowIndex}
-        column={column}
-        value={value}
-        onChange={setValue}
-        onBlur={onBlur}
-        onKeyDown={onKeyDown}
-        {...props} />
-  )
-};
-
-const CellWrapper = ({
-  defaultEditComponents = {},
-  readOnly,
-  value,
-  onChange,
-  columns,
-  selectedCells,
-  onSelectedCellsChange,
-  rowIndex,
-  columnIndex,
-  editingCell,
-  onEditingCellChange,
-  totalRows,
-  totalColumns
-}) => {
-  const { hoverRow, onHoverRowChange } = useContext(CellsRowContext);
-
-  const column = columns[columnIndex];
-  const rowValue = value[rowIndex];
-  const curValue = get(rowValue, column.valuePath);
-  const editing = !readOnly && editingCell && (editingCell.row === rowIndex && editingCell.column === columnIndex);
-  const selectedRow = selectedCells.some(selectedCell => selectedCell.row === rowIndex);
-  const selectedCell = selectedCells.some(selectedCell => selectedCell.row === rowIndex && selectedCell.column === columnIndex);
-
-  const onMouseEnter = useCallback(() => onHoverRowChange(true), [onHoverRowChange]);
-  const onMouseLeave = useCallback(() => onHoverRowChange(false), [onHoverRowChange]);
-
-  const handleSelect = useCallback(({ ctrlKey, shiftKey }) => {
-    onSelectedCellsChange(selectedCells => {
-      const lastSelection = selectedCells[selectedCells.length - 1];
-      const curSelection = { row: rowIndex, column: columnIndex };
-
-      let nextSelection;
-      if (ctrlKey && lastSelection) {
-        nextSelection = [...selectedCells, curSelection];
-      } else if (shiftKey && lastSelection) {
-        const startRowIndex = selectedCells.reduce((prev, selection) => Math.min(prev, selection.row, curSelection.row), value.length - 1);
-        const endRowIndex = selectedCells.reduce((prev, selection) => Math.max(prev, selection.row, curSelection.row), 0);
-        nextSelection = [];
-        for (let i = startRowIndex; i <= endRowIndex; i++) {
-          nextSelection.push({ row: i });
-        }
-      } else {
-        nextSelection = [curSelection];
-      }      
-      return nextSelection;
-    });
-  }, [columnIndex, rowIndex, onSelectedCellsChange, value.length]);
-
-  const startEditCell = useCallback(() => {
-    onEditingCellChange({ row: rowIndex, column: columnIndex });
-  }, [onEditingCellChange, rowIndex, columnIndex]);
-
-  const baseComponentProps = {
-    rowIndex,
-    columnIndex,
-    column,
-    value: curValue,
-    onMouseDown: handleSelect,
-    onMouseEnter: onMouseEnter,
-    onMouseLeave: onMouseLeave,
-    className: classNames(
-      'cell',
-      column.type || 'string',
-      {
-        'hover-row': hoverRow,
-        'selected-row': selectedRow,
-        'selected-cell': selectedCell
-      }
-    )
-  };
-
-  const EditComponent = column.EditComponent || defaultEditComponents[column.type || 'string'];
-  const ValueComponent = column.Component || Cell;
-
-  let Component, componentProps;
-  if (editing && EditComponent) {
-    Component = EditCellWrapper;
-    componentProps = {
-      ...baseComponentProps,
-      Component: EditComponent,
-      className: classNames(baseComponentProps.className, 'edit'),
-      onChange,
-      onEditingCellChange,
-      onSelectedCellsChange,
-      totalRows,
-      totalColumns
-    };
-  } else {
-    Component = ValueComponent;
-    componentProps = {
-      ...baseComponentProps,
-      onDoubleClick: startEditCell
-    };
-  }
-
-  return <Component {...componentProps} />;
-};
-
-const Cells = React.memo(({
-  visibleRows,
-  visibleColumns,
-  ...props
-}) => {
-  return visibleRows.map(rowIndex => {
-    const columnsElements = (
-      <CellsRow key={rowIndex}>
-        {visibleColumns.map(columnIndex => {
-          return (
-            <CellWrapper
-                key={`${rowIndex}-${columnIndex}`}
-                rowIndex={rowIndex}
-                columnIndex={columnIndex}
-                {...props} />
-          );
-        })}
-      </CellsRow>
-    );
-    return columnsElements;
-  }, [])
 });
 
 const FooterItem = ({ value, column, style }) => {
